ball_physics: match 13 more functions (36/40)
- Retype g_Ball.ballZoneWhenCaught as s8: it is stored as -1 and read
  signed. Adjust the fielder, fielder_ai and runner readers so those
  units are unchanged.
- Use the typed stadiumObjectCollision from stadium_framework.h instead
  of a local byte-array extern.
- Reuse foulBall() for the duplicated foul-ball blocks and drop the
  do/while(0) break idiom from processLandedBallBouncing and
  processBallInAir_Landed.
- Reset family, setDefaultInMemBall and initBallAndGameStateOnLoad: read
  FrameCountOfEntireGame once, write the RNG terms shift-first and move
  the independent stores after the arithmetic.
- estimateWhereBallWillHitWall, calculateImplicationsOfTheHitTrajectory
  (via an inlined closest-future-frame search), warioWaluStarHit,
  fairOrFoulBall, setLiveBallVariablesAfterContact, liveBallHitPhysics
  and chompCollision_processStarHitVariables now match.

Still below 100%: estimateAndSetFutureCoords, ballCollisionLogic,
classifyHitTrajectoryOrHitAnimRelated, handleBallBounceAndRoll.

+++ src/game/baserunning/runner.c
@@ -4176,7 +4176,7 @@ void running_UpdateTrackingValues_BasesEarned(void) {
     if (g_Runners[0].baseNumberEarned_NotIncludingFieldersChoice == -1) {
         u8 ballState = g_Ball.ballState;
         u8 ballZone = g_Ball.ballZoneAwayFromHome;
-        u8 caughtZone = g_Ball.ballZoneWhenCaught;
+        u32 caughtZone = (u8)g_Ball.ballZoneWhenCaught;
         u8 playOver = g_FieldingLogic.playOverInd;
 
         runner = g_Runners;
